requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-notifications